relocate coordinate mapper remove coordinate mapper interface update api location datasource
