+++ src/main/kotlin/data/location/datasource/ApiLocationDataSource.kt
@@ -2,16 +2,18 @@ package data.location.datasource
 
 import data.location.model.CoordinateDto
 import io.ktor.client.*
-import io.ktor.client.engine.cio.*
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
 import kotlinx.serialization.json.Json
 
-class ApiLocationDataSource: LocationDataSource {
+class ApiLocationDataSource(
+    private val client: HttpClient,
+): LocationDataSource {
     override suspend fun getLocationByCityAndCountry(cityName: String, country: String): CoordinateDto {
-        val client = HttpClient(CIO)
-        val response = client.get("https://geocoding-api.open-meteo.com/v1/search?name=$cityName,$country")
-        val data = Json.decodeFromString<CoordinateDto>(response.bodyAsText())
-        return data
+        val response = client.get{
+            url("https://geocoding-api.open-meteo.com/v1/search")
+            parameter("name", "$cityName,$country")
+        }
+        return Json.decodeFromString<CoordinateDto>(response.bodyAsText())
     }
 }
