+++ src/main/kotlin/data/location/mapper/CoordinateMapper.kt
@@ -1,11 +1,11 @@
-package data.location.repository.mapper
+package data.location.mapper
 
 import data.location.model.CoordinateDto
 import logic.exception.NoLocationFoundException
 import logic.model.Coordinate
 
-class CoordinateMapperImpl : CoordinateMapper {
-    override fun mapDtoToCoordinate(coordinateDto: CoordinateDto): Coordinate {
+class CoordinateMapper {
+    fun mapDtoToCoordinate(coordinateDto: CoordinateDto): Coordinate {
         return coordinateDto.citiesCoordinates.firstOrNull()
             ?.let { Coordinate(it.latitude, it.longitude)}
             ?: throw NoLocationFoundException()
