--- src/main/kotlin/data/location/repository/mapper/CoordinateMapper.kt
@@ -1,8 +1,0 @@
-package data.location.repository.mapper
-
-import data.location.model.CoordinateDto
-import logic.model.Coordinate
-
-interface CoordinateMapper {
-    fun mapDtoToCoordinate(coordinateDto: CoordinateDto): Coordinate
-}
